Autogenerate 2D nontile permutations
Added generate_codes.sh under 2D/nonTiled. Now we can autogenerate the C code for each of the permutations under nonTiled.
Next step - add the permutations for tiled analogous to the way we've done nonTiled. And then also use the generate_code.sh script to also loop over those file too.

Next next step - do this same thing for the rest of the codes.

final step - once we have all of the binaries for each type of code, then we can write a script to invoke each with parameters and save output somewhere.